requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
streamlit>=1.30.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-anthropic>=0.1.0
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
